EMPIREDB-103 part 1: New base class DBDDLGenerator and implmentations for HSql, SqlServer and Oracle. Other driver will follow
git-svn-id: https://svn.apache.org/repos/asf/incubator/empire-db/trunk@1180631 13f79535-47bb-0310-9956-ffa450edef68

+++ empire-db/src/main/java/org/apache/empire/data/DataType.java
@@ -61,9 +61,9 @@ public enum DataType
     CHAR, //         = 6;
     
     /**
-     * Double precision floating point value
+     * floating point value
      */
-    DOUBLE, //       = 7;
+    FLOAT, //        = 7;
     
     /**
      * Decimal numeric value (size indicates scale and precision)
